14_Authentications/role_based_access/requirements.txt
fastapi>=0.115,<0.116
uvicorn[standard]>=0.30,<0.31
PyJWT>=2.9,<3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic>=2.7,<3
typing-extensions>=4.12,<5
python-dotenv>=1.0,<2
email-validator>=2,<3
